flattened prover module structure
