+++ prover/src/domain.rs
