+++ prover/src/trace/execution_trace.rs
