+++ prover/src/trace/mod.rs
