+++ prover/src/trace/poly_table.rs
