+++ prover/src/trace/trace_table.rs
